requirements.txt
flask
python-dotenv
flask-restful
flask-sqlalchemy
flask-marshmallow
Faker